refactor: split apollo-auth into three modules for Maven Central
Split the monolithic apollo-auth module into three separate modules:

- apollo-core: Type classes (HasId, HasEmail, HasPassword, Hashable)
  and service traits (UserService, SessionService, etc.)
  Dependencies: cats, cats-effect only

- apollo-doobie: Doobie-based implementations of the service traits
  Dependencies: apollo-core, doobie

- apollo-http4s: Http4s routes, middleware, and Twirl templates
  Dependencies: apollo-core, http4s, twirl, upickle

This modular structure allows users to:
- Use just the core abstractions and implement their own backends
- Mix and match database implementations (Doobie, Skunk, etc.)
- Use the Http4s layer independently of the database layer

Also removes MSSQL support in favor of PostgreSQL only.
