--- apollo-auth/src/db/mssql/000-create-users.sql
@@ -1,29 +1,0 @@
-USE [webapp]
-GO
-
-SET ANSI_NULLS ON
-GO
-
-SET QUOTED_IDENTIFIER ON
-GO
-
-CREATE TABLE [dbo].[users](
-	[id] [uniqueidentifier] NOT NULL,
-	[name] [nvarchar](320) NOT NULL,
-	[email] [nvarchar](320) NOT NULL,
-	[password] [nvarchar](161) NOT NULL,
-	[confirmed_at] [datetimeoffset](7) NULL,
-	[created_at] [datetimeoffset](7) NOT NULL,
-PRIMARY KEY CLUSTERED
-(
-	[id] ASC
-)WITH (PAD_INDEX = OFF, STATISTICS_NORECOMPUTE = OFF, IGNORE_DUP_KEY = OFF, ALLOW_ROW_LOCKS = ON, ALLOW_PAGE_LOCKS = ON, OPTIMIZE_FOR_SEQUENTIAL_KEY = OFF) ON [PRIMARY],
-UNIQUE NONCLUSTERED
-(
-	[email] ASC
-)WITH (PAD_INDEX = OFF, STATISTICS_NORECOMPUTE = OFF, IGNORE_DUP_KEY = OFF, ALLOW_ROW_LOCKS = ON, ALLOW_PAGE_LOCKS = ON, OPTIMIZE_FOR_SEQUENTIAL_KEY = OFF) ON [PRIMARY]
-) ON [PRIMARY]
-GO
-
-ALTER TABLE [dbo].[users] ADD  DEFAULT (lower(newid())) FOR [id]
-GO
