--- apollo-auth/views/layout.scala.html
@@ -1,25 +1,0 @@
-@(content: Html)
-<!DOCTYPE html>
-<html>
-<head>
-    <meta charset="utf-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1">
-    <link href="/normalize.css/8.0.1/normalize.css" rel="stylesheet">
-    <link href="/bootstrap/5.3.3/dist/css/bootstrap.min.css" rel="stylesheet">
-    <script type="text/javascript" src="/jquery/4.0.0-beta.2/dist/jquery.min.js"></script>
-    <script type="text/javascript" src="/bootstrap/5.3.3/dist/js/bootstrap.min.js"></script>
-</head>
-<body class="bg-body-tertiary">
-<div class="container min-vh-100 d-flex align-items-center justify-content-center py-5">
-    <div class="row justify-content-center w-100">
-        <div class="col-12 col-sm-10 col-md-8 col-lg-6 col-xl-5">
-            <div class="card border-0 shadow-none rounded-3 bg-white">
-                <div class="card-body p-4 p-md-5">
-                    @content
-                </div>
-            </div>
-        </div>
-    </div>
-</div>
-</body>
-</html>
