+++ apollo-doobie/src/db/pg/002-create-sessions.sql
