+++ apollo-doobie/src/db/pg/initialize-database.bat
