+++ apollo-http4s/views/changePasswordForm.scala.html
@@ -1,0 +1,36 @@
+@(csrfToken: String, flash: Option[String] = None)
+
+@layout {
+<h2 id="title">Password reset</h2>
+<p>Change your password</p>
+
+<div style="max-width: 26rem">
+    @flash.map { message =>
+    <div class="alert alert-danger" role="alert">@message.</div>
+    }
+
+    <div class="tab-content">
+        <div id="pills-login" role="tabpanel">
+            <form method="POST">
+                <input type="hidden" name="_csrf" value="@csrfToken">
+
+                <div class="form-outline mb-4">
+                    <input class="form-control" type="password" name="password"
+                           id="resetPassword" placeholder="Password">
+                </div>
+
+                <div class="form-outline mb-4">
+                    <input class="form-control" type="password" name="confirmPassword"
+                           id="resetConfirmPassword" placeholder="Confirm password">
+                </div>
+
+                <div class="d-grid mx-auto">
+                    <button class="btn btn-primary btn-block mb-4" type="submit">
+                        Change password
+                    </button>
+                </div>
+            </form>
+        </div>
+    </div>
+</div>
+}
