+++ apollo-http4s/views/resetRequestForm.scala.html
@@ -1,0 +1,35 @@
+@(csrfToken: String, flash: Option[(String, String)] = None)
+
+@layout {
+<h2 id="title">Password reset</h2>
+<p>Request a password reset</p>
+
+<div style="max-width: 26rem">
+    @flash.map { case (cssClass, message) =>
+    <div class="alert @cssClass" role="alert">@message.</div>
+    }
+
+    <div class="tab-content">
+        <div id="pills-login" role="tabpanel">
+            <form action="/reset" method="POST">
+                <input type="hidden" name="_csrf" value="@csrfToken">
+
+                <div class="form-outline mb-4">
+                    <input class="form-control" type="email" name="email"
+                           id="loginName" placeholder="Email" required>
+                </div>
+
+                <div class="d-grid mx-auto">
+                    <button class="btn btn-primary btn-block mb-4" type="submit">
+                        Submit request
+                    </button>
+                </div>
+
+                <div class="text-center">
+                    <a href="/" class="btn btn-link p-0 me-2">Back to login</a>
+                </div>
+            </form>
+        </div>
+    </div>
+</div>
+}
